@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "engines": {
-    "node": "16.14.2"
+    "node": "22.14.0"
   },
   "dependencies": {
     "@headlessui/react": "^1.7.15",
@@ -19,7 +19,7 @@
     "@types/react-bootstrap": "^0.32.22",
     "@types/react-router-dom": "^5.3.3",
     "antd": "^4.7.3",
-    "bootstrap": "^4.6.1",
+    "bootstrap": "^5.3.3",
     "classnames": "^2.2.6",
     "date-fns": "^3.5.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
@@ -27,9 +27,8 @@
     "jquery": "^3.6.0",
     "jwt-decode": "^2.2.0",
     "normalize.css": "^8.0.1",
-    "postcss": "^8.4.26",
     "postcss-loader": "^7.0.0",
-    "react": "^18.0.0",
+    "react": "^18.3.1",
     "react-alert": "^7.0.0",
     "react-alert-template-basic": "^1.0.0",
     "react-anchor-link-smooth-scroll": "^1.0.12",
@@ -40,7 +39,7 @@
     "react-bootstrap-table2-toolkit": "^2.1.3",
     "react-datepicker": "^2.16.0",
     "react-day-picker": "^8.10.0",
-    "react-dom": "^18.0.0",
+    "react-dom": "^18.3.1",
     "react-drag-drop-files": "^2.1.14",
     "react-dropzone": "^11.4.2",
     "react-dropzone-uploader": "^2.11.0",
@@ -58,22 +57,17 @@
     "react-redirect": "^1.0.0",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "^5.0.1",
     "react-select": "^3.1.0",
     "react-stepper-horizontal": "^1.0.11",
     "react-switch": "^5.0.1",
     "react-test-renderer": "^18.0.0",
     "react-text-mask": "^5.4.3",
     "react-uuid": "^1.0.2",
-    "sass": "^1.63.3",
     "typescript": "^4.6.3",
-    "valid-url": "^1.0.9",
-    "web-vitals": "^2.1.4"
+    "valid-url": "^1.0.9"
   },
   "devDependencies": {
     "@testing-library/dom": "^7.29.0",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.0.1",
     "@testing-library/user-event": "^13.5.0",
     "@types/enzyme": "3.10.8",
     "@types/enzyme-adapter-react-16": "1.0.6",
@@ -82,9 +76,11 @@
     "@types/jquery-mask-plugin": "^1.14.3",
     "@types/react-mailchimp-subscribe": "^2.1.4",
     "@typescript-eslint/parser": "^2.34.0",
-    "autoprefixer": "^6.7.2",
+    "@vitejs/plugin-react": "^4.3.4",
+    "autoprefixer": "^10.4.21",
     "babel-preset-react-app": "^2.2.0",
     "babel-runtime": "^6.20.0",
+    "browserslist": "^4.24.4",
     "case-sensitive-paths-webpack-plugin": "1.1.4",
     "chalk": "4.1.0",
     "connect-history-api-fallback": "1.3.0",
@@ -108,24 +104,26 @@
     "msw": "^0.26.2",
     "node-hot-loader": "^1.20.0",
     "object-assign": "4.1.1",
+    "postcss": "^8.5.3",
     "prettier-eslint": "^12.0.0",
     "promise": "7.1.1",
     "react-cosmos": "^5.5.3",
     "react-dev-utils": "^12.0.1",
     "react-hot-loader": "^4.13.0",
     "redirect": "^0.2.0",
+    "sass": "^1.86.0",
     "style-loader": "0.13.1",
     "stylelint": "^12.0.1",
-    "tailwindcss": "^3.3.3",
+    "tailwindcss": "^3.4.17",
     "ts-jest": "27.1.4",
     "url-loader": "^2.2.0",
+    "vite": "^6.2.3",
     "whatwg-fetch": "2.0.2"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "cross-env INLINE_RUNTIME_CHUNK=false DISABLE_ESLINT_PLUGIN=true TSC_COMPILE_ON_ERROR=true ESLINT_NO_DEV_ERRORS=true GENERATE_SOURCEMAP=false react-scripts build",
-    "test": "react-scripts test -u",
-    "eject": "react-scripts eject",
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
     "lint:scss": "stylelint 'src/**/*.scss' --syntax scss",
     "lint:scss:fix": "stylelint 'src/**/*.scss' --syntax scss --fix",
     "lint:ts": "eslint . --ext .ts,.tsx",
