@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { Component, useState } from 'react';
 import { withAlert } from 'react-alert';
 import { Link } from 'react-router-dom';
 
@@ -23,86 +23,15 @@ interface Props {
 
 interface State {}
 
-export class Header extends Component<Props, State, {}> {
-  constructor(props: Props) {
-    super(props);
-    this.state = {};
-  }
+// We extend React.Component with Props & State
+function Header({ logIn, logOut, isLoggedIn, role, alert }: Props) {
+  const [isOpen, setIsOpen] = useState(false);
 
-  render() {
-    const { isLoggedIn, role, logOut } = this.props;
-    if (isLoggedIn) {
-      return (
-        <nav className="navbar navbar-expand-lg navbar-dark sticky-top navbar-custom">
-          <div className="container">
-            <Link className="pr-3" to="/home">
-              <img
-                alt="Logo"
-                src={Logo}
-                width={logoSize}
-                height={logoSize}
-                className="d-inline-block"
-              />
-            </Link>
-            <Link className="navbar-brand" to="/home">
-              Keep.id
-            </Link>
-            <button
-              className="navbar-toggler"
-              type="button"
-              data-toggle="collapse"
-              data-target="#navbarToggleLoggedIn"
-              aria-controls="navbarToggleLoggedIn"
-              aria-expanded="false"
-              aria-label="Toggle navigation"
-            >
-              <span className="navbar-toggler-icon" />
-            </button>
-            <div
-              className="navbar-collapse collapse w-100 order-3 dual-collapse2"
-              id="navbarToggleLoggedIn"
-            >
-              <ul className="navbar-nav ml-auto">
-                <li className="nav-item col-med-2 my-1 flex-fill mr-2">
-                  <Link className="nav-link" to="/">
-                    {role === Role.Admin ||
-                    role === Role.Director ||
-                    role === Role.Worker
-                      ? 'My Clients'
-                      : 'Home'}
-                  </Link>
-                </li>
-                {(role === Role.Admin || role === Role.Director) && (
-                  <li className="nav-item col-med-2 my-1 flex-fill mr-2">
-                    <Link className="nav-link" to="/admin-panel">
-                      Admin Panel
-                    </Link>
-                  </li>
-                )}
-                <li className="nav-item col-med-2 my-1 flex-fill mr-2">
-                  <Link className="nav-link" to="/settings">
-                    My Account Settings
-                  </Link>
-                </li>
-                {(role === Role.Admin || role === Role.Director) && (
-                  <li className="nav-item col-med-2 my-1 flex-fill mr-2">
-                    <Link className="nav-link" to="/my-organization">
-                      My Organization
-                    </Link>
-                  </li>
-                )}
-                <div className="my-1 flex-fill">
-                  <Logout logOut={logOut} />
-                </div>
-              </ul>
-            </div>
-          </div>
-        </nav>
-      );
-    }
+  if (isLoggedIn) {
     return (
-      <nav className="navbar navbar-expand-lg navbar-dark sticky-top navbar-custom">
+      <nav className="navbar navbar-dark sticky-top">
         <div className="container">
+          {/* Logo and Brand */}
           <Link className="pr-3" to="/home">
             <img
               alt="Logo"
@@ -115,61 +44,174 @@ export class Header extends Component<Props, State, {}> {
           <Link className="navbar-brand" to="/home">
             Keep.id
           </Link>
+
+          {/* Toggler */}
           <button
             className="navbar-toggler"
             type="button"
             data-toggle="collapse"
-            data-target="#navbarToggle"
-            aria-controls="navbarToggle"
+            data-target="#navbarToggleLoggedIn"
+            aria-controls="navbarToggleLoggedIn"
             aria-expanded="false"
             aria-label="Toggle navigation"
           >
             <span className="navbar-toggler-icon" />
           </button>
 
-          <div className="collapse navbar-collapse" id="navbarToggle">
-            <ul className="navbar-nav ml-auto mt-2 mt-lg-0 ">
-              <li className="nav-item my-1 mr-2 ml-2">
+          {/* Collapsible Section */}
+          <div
+            className="collapse navbar-collapse w-100 order-3"
+            id="navbarToggleLoggedIn"
+          >
+            <ul className="navbar-nav ml-auto">
+              <li className="nav-item mr-2 my-1">
                 <Link className="nav-link" to="/">
-                  Home
+                  {role === Role.Admin ||
+                  role === Role.Director ||
+                  role === Role.Worker
+                    ? 'My Clients'
+                    : 'Home'}
                 </Link>
               </li>
-              <li className="nav-item my-1 mr-2 ml-2">
-                <Link className="nav-link" to="/find-organizations">
-                  Find Organizations
+
+              {(role === Role.Admin || role === Role.Director) && (
+                <li className="nav-item mr-2 my-1">
+                  <Link className="nav-link" to="/admin-panel">
+                    Admin Panel
+                  </Link>
+                </li>
+              )}
+
+              <li className="nav-item mr-2 my-1">
+                <Link className="nav-link" to="/settings">
+                  My Account Settings
                 </Link>
               </li>
-              <li className="nav-item my-1 mr-3 ml-2">
-                <a className="nav-link" href="https://team.keep.id">
-                  About
-                </a>
+
+              {(role === Role.Admin || role === Role.Director) && (
+                <li className="nav-item mr-2 my-1">
+                  <Link className="nav-link" to="/my-organization">
+                    My Organization
+                  </Link>
+                </li>
+              )}
+
+              {/* Logout button in its own li */}
+              <li className="nav-item my-1">
+                <Logout logOut={logOut} />
               </li>
             </ul>
-            <div className="my-1">
-              <Link to="/login">
-                <button
-                  type="submit"
-                  className="btn btn-primary outline-white w-40 mr-2"
-                >
-                  Log in
-                </button>
-              </Link>
-            </div>
-            <div className="my-1">
-              <Link to="/signup-branch">
-                <button
-                  type="submit"
-                  className="btn btn-outline-secondary w-40 mr-2"
-                >
-                  Sign Up
-                </button>
-              </Link>
-            </div>
           </div>
         </div>
       </nav>
     );
   }
+
+  // If not logged in
+  return (
+    <nav className="tw-bg-gray-800 tw-text-white tw-w-full tw-sticky tw-top-0">
+    {/* Main container: brand and hamburger */}
+    <div className="tw-flex tw-items-center tw-justify-between tw-py-3 tw-px-4">
+      {/* Left side: Logo + Brand */}
+      <div className="tw-flex tw-items-center">
+        <Link to="/home" className="tw-flex tw-items-center tw-pr-3">
+          <img
+            alt="Logo"
+            src={Logo}
+            width={logoSize}
+            height={logoSize}
+          />
+        </Link>
+        <Link to="/home" className="tw-text-xl tw-font-semibold">
+          Keep.id
+        </Link>
+      </div>
+
+      {/* Hamburger button (only visible on small screens) */}
+      <div className="sm:tw-hidden">
+        <button
+          onClick={() => setIsOpen(!isOpen)}
+          className="tw-text-white tw-outline-none tw-focus:outline-none"
+        >
+          <svg
+            className="tw-w-6 tw-h-6"
+            fill="none"
+            stroke="currentColor"
+            viewBox="0 0 24 24"
+            xmlns="http://www.w3.org/2000/svg"
+          >
+            {isOpen ? (
+              // X (Close) icon
+              <path
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                strokeWidth="2"
+                d="M6 18L18 6M6 6l12 12"
+              />
+            ) : (
+              // Hamburger icon
+              <path
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                strokeWidth="2"
+                d="M4 6h16M4 12h16M4 18h16"
+              />
+            )}
+          </svg>
+        </button>
+      </div>
+    </div>
+
+    {/* Collapsible menu: hidden on small screens unless isOpen is true */}
+    <div
+      className={`tw-bg-gray-800 sm:tw-flex sm:tw-items-center sm:tw-justify-end 
+        ${isOpen ? 'tw-block' : 'tw-hidden'}`}
+    >
+      <ul className="sm:tw-flex sm:tw-space-x-4">
+        <li className="tw-my-2 sm:tw-my-0">
+          <Link to="/" className="tw-block tw-py-2 tw-px-4 hover:tw-bg-gray-700">
+            Home
+          </Link>
+        </li>
+        <li className="tw-my-2 sm:tw-my-0">
+          <Link
+            to="/find-organizations"
+            className="tw-block tw-py-2 tw-px-4 hover:tw-bg-gray-700"
+          >
+            Find Organizations
+          </Link>
+        </li>
+        <li className="tw-my-2 sm:tw-my-0">
+          <a
+            href="https://team.keep.id"
+            className="tw-block tw-py-2 tw-px-4 hover:tw-bg-gray-700"
+          >
+            About
+          </a>
+        </li>
+      </ul>
+
+      <div className="tw-flex tw-flex-col sm:tw-flex-row tw-items-center tw-mb-2 sm:tw-mb-0 sm:tw-ml-4">
+        <Link to="/login" className="tw-mb-2 sm:tw-mb-0 sm:tw-mr-2">
+          <button
+            type="button"
+            className="tw-bg-blue-600 tw-text-white tw-py-2 tw-px-4 tw-rounded hover:tw-bg-blue-700"
+          >
+            Log in
+          </button>
+        </Link>
+        <Link to="/signup-branch">
+          <button
+            type="button"
+            className="tw-border tw-border-gray-300 tw-text-white tw-py-2 tw-px-4 tw-rounded hover:tw-bg-gray-700"
+          >
+            Sign Up
+          </button>
+        </Link>
+      </div>
+    </div>
+  </nav>
+  );
 }
 
 export default withAlert()(Header);
