@@ -4,7 +4,6 @@ import AlertTemplate from 'react-alert-template-basic';
 import { IntlProvider } from 'react-intl';
 import { createRoot } from 'react-dom/client';
 
-import reportWebVitals from './reportWebVitals';
 import App from './App';
 
 const options = {
@@ -17,7 +16,7 @@ const options = {
     zIndex: 99999,
   },
 };
-const root = createRoot(document.getElementById('root'));
+const root = createRoot(document.getElementById('root'))
 root.render(
   <Provider
     template={AlertTemplate}
@@ -28,9 +27,4 @@ root.render(
       <App />
     </IntlProvider>
   </Provider>
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+)
