@@ -1,14 +1,13 @@
-const getServerURL = (): string => {
-  if (process.env.REACT_APP_KEEPID_BUILD_ENV === 'staging') {
-    return 'https://staged.keep.id';
+const getServerURL = () => {
+  const currentMode = import.meta.env.MODE;
+
+  if (currentMode === 'staging') {
+    return 'https://staging.keep.id';
   }
-  if (process.env.NODE_ENV === 'production') {
+  if (currentMode === 'production') {
     return 'https://server.keep.id';
   }
-  if (process.env.NODE_ENV === 'staging') {
-    return 'https://staged.keep.id';
-  }
-  return 'http://localhost:7000';
+  return 'http://localhost:7001';
 };
 
 export default getServerURL;
